issue #23: 軽微なリファクタリング

+++ src/optimization/runner/la40.rs
@@ -30,8 +30,8 @@ fn initialize_chromosome(num_job: u16, num_actor: u16) -> types::Chromosome {
 }
 
 pub fn run(job_master: JobMaster) {
-    let num_job: u16 = job_master.job_size as u16;
-    let num_actor: u16 = job_master.machine_series_size as u16;
+    let num_job: u16 = job_master.job_size;
+    let num_actor: u16 = job_master.machine_series_size;
     let mut chromosome: types::Chromosome =
         initialize_chromosome(num_job, num_actor);
